+++ ts/bot/botcontents.ts
