+++ ts/bot/oxgame/oxgametype.ts
